item lock db query optimization
